requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.8.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
watchgod==0.7
websockets==10.1
